+++ optuna/samplers/__init__.py
@@ -1,7 +1,7 @@
 from optuna.samplers._search_space import intersection_search_space  # NOQA
 from optuna.samplers._search_space import IntersectionSearchSpace  # NOQA
-from optuna.samplers.base import BaseSampler  # NOQA
-from optuna.samplers.cmaes import CmaEsSampler  # NOQA
-from optuna.samplers.grid import GridSampler  # NOQA
-from optuna.samplers.random import RandomSampler  # NOQA
-from optuna.samplers.tpe import TPESampler  # NOQA
+from optuna.samplers._base import BaseSampler  # NOQA
+from optuna.samplers._cmaes import CmaEsSampler  # NOQA
+from optuna.samplers._grid import GridSampler  # NOQA
+from optuna.samplers._random import RandomSampler  # NOQA
+from optuna.samplers._tpe import TPESampler  # NOQA
