+++ optuna/samplers/_tpe/__init__.py
@@ -1,0 +1,1 @@
+from optuna.samplers._tpe.sampler import TPESampler  # NOQA
