--- optuna/samplers/tpe/__init__.py
@@ -1,1 +1,0 @@
-from optuna.samplers.tpe.sampler import TPESampler  # NOQA
