+++ tests/samplers_tests/tpe_tests/test_parzen_estimator.py
@@ -3,8 +3,8 @@ import itertools
 import numpy as np
 import pytest
 
-from optuna.samplers.tpe.parzen_estimator import _ParzenEstimator
-from optuna.samplers.tpe.sampler import default_weights
+from optuna.samplers._tpe.parzen_estimator import _ParzenEstimator
+from optuna.samplers._tpe.sampler import default_weights
 from optuna import type_checking
 
 if type_checking.TYPE_CHECKING:
